refactor code, move internal functions to internal
