+++ internal/error/error.go
@@ -1,8 +1,8 @@
-package jwt
+package error
 
 import "fmt"
 
-func appendError(e1, e2 error) error {
+func AppendError(e1, e2 error) error {
 	if e1 == nil {
 		return e2
 	}
@@ -12,7 +12,7 @@ func appendError(e1, e2 error) error {
 	return fmt.Errorf("%s; %s", e1.Error(), e2.Error())
 }
 
-func prefixError(e error, p string) error {
+func PrefixError(e error, p string) error {
 	if e != nil {
 		return fmt.Errorf("%s %s", p, e.Error())
 	}
