+++ internal/stdin/read.go
