+++ internal/stdin/read_test.go
@@ -133,7 +133,7 @@ func TestRead(t *testing.T) { // simple test to make sure read is actually using
 	defer func() { os.Stdin = save }()
 	os.Stdin = r // temporarily mangle stdin with our own pipe for testing
 
-	actual := Read()
+	actual := read()
 
 	if actual != testVal {
 		t.Logf("read value doesn't appear to be from os.Stdin!")
