+++ internal/stdin/test.txt
